Code/pi_code/requirements.txt
# Pi5 Hardware Monitoring System Dependencies

# Hardware drivers
gpiozero>=2.0
mpu6050-raspberrypi>=1.1.1

# Supabase communication
requests>=2.31.0
python-dotenv>=1.0.0  # For environment variable management (optional)

# Other
colorama>=0.4.6  # Colored console output (optional)
